requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
discord.py==0.16.12
idna==2.7
idna-ssl==1.1.0
inflection==0.3.1
multidict==4.4.2
websockets==3.4
yarl==1.2.6
